find-presentation-translators: refactor the code
Changes are mostly made for readibility. One notable change is that we
call eagerly position for x-precedence (2nd test) while previously we
have called it only when we got to the second clause. It does not
matter much because results are cached.

We curb nest macro from uiop and use it to make local function
translator-lessp denser (still more readable thanks to reduced
indentation changes).

Despite the comment above the local function, floor was called without
the second argument. Since priority is always integer it was basically
identity - it is fixed according to what is said in the comment.