window-configuration-event fix
In window-configuration-event the position: x, y and the size: width,
height are stored in native coordinates of the parent of the sheet of
the event.

For this reason I rename the reader accessor of width and height slots
as WINDOW-CONFIGURATION-EVENT-NATIVE-{WIDTH,HEIGHT}

The methods WINDOW-CONFIGURATION-EVENT-{X,Y,WIDTH,HEIGHT} transform
from parent native coordinate to parent sheet coordinate.

The methods WINDOW-CONFIGURATION-EVENT-{X,Y} were wrong (they used the
sheet-native-transformation of the sheet and not of the parent and use
transform-position instead of untransform-position) so I correct them.

I also write the new WINDOW-CONFIGURATION-EVENT-{WIDTH,HEIGHT} methods.