sheet-transformation for top-level-sheet-pane
This commit revert the changes of commit a894385 in PR #691 that
solved the bug of issue #499 about the sliding to the left, of the
content of a window protruded to the left of the screen.

In that commit
(handle-event top-level-sheet-pane window-configuration-event)
method was changed in the following way:

1. WINDOW-CONFIGURATION-EVENT-{X,Y} were changed with
   WINDOW-CONFIGURATION-EVENT-NATIVE-{X,Y}, but the first is right
   because the sheet-transformation isn't expressed in native
   coordinates. The problem was in the implementation of
   WINDOW-CONFIGURATION-EVENT-{X,Y} that is solved in the previous
   commit. So now is possible to use WINDOW-CONFIGURATION-EVENT-{X,Y}
   without bugs.

2. The method avoided to set as sheet-transformation a translation
   with negative dx or dy. This fix the bug of the sliding content but
   in this way the sheet-transformation is wrong (issue #769): it
   isn't the transformation that converts coordinates in the sheet
   sheet's coordinate system into coordinates in its parent's coordinate
   system.

As noted in the message of commit a894385 the bug of the sliding content
is due to the fact that the translation with negative offset were
recorded two times: one as sheet-transformation and another one as
sheet-native-transformation.

Here I change function UPDATE-MIRROR-GEOMETRY in such a way that the
translation with negative offset is recorded only as
sheet-transformation and no more as sheet-native-transformation.

For this, at the begining of the function, I change the calculation
of sheet-region-in-native-parent for the top-level-sheet-panes that
have a graft as parent. For this kind of panes it is not correct to take the
intersection between the sheet-region and the sheet-region of the
parent because the sheet native region can start outside the parent
native region.

Fixes #769 #499