presentation-translator: introduce define-selection-translator macro
Selection translators are simplified presentation-translators. They
are always bound to a gesture :select, tester is always definitive and
does not depend on arguments (context-type frame event window x y)
which can't appear on the arglist. That means that selection
translators may be used in a processing context which doesn't involve
frame and a pointer (i.e in selection/clipboard protocol).

cleanup: presentation-translators: detabify and add warning

Warning is added to prevent complaining users (see #778 for the
issue). Detabify and small code cleanup follows. Also we fix scope and
remove some clutter from add-translator.