stream-input: separate input-buffer from event-queue
We remove invalid assumption that input-buffer is same thing as
event-queue for standard-extended-input-stream. To do that we split
the initargs and slots between standard-sheet-input-mixin and
standard-input-stream (along with appropriate accessors).

This change is incomplete, because stream reading operations still
assume that the input-buffer is an event-queue what leads to debugger
popping all the time. To make this commit atomically complete we
replace passages in the code which make this invalid assumption with
event-sheet-queue access (and add appropriate comment that this is
invalid). This will be fixed in upcoming commit. Software after this
change will work as before except that SEIS will signal unbound-slot
when we try to access STREAM-INPUT-BUFFER from the user code (access
from McCLIM has been replaced).

- add :type vector to the input-buffer slots
- untie input-buffer and event-queue
- remove stream-input-buffer references from sheet-input
- replace stream-input-buffer with sheet-event-queue where appropriate
- minor refactor of some code passages (stylistic)