Fix (setf output-record-position) for draw-bezier-design.
When setf'ing the position of a bezier design, the coordinates stay unmodified
while a standard-transformation is used to store the delta-x/y translation from
the original position to the new one.
The fault with this method was that the translations were overwritten each time,
with the result (in an interactive setting) that the bezier-design wiggles
around by a few pixels for each pointer-motion event, instead of adding the
translations on top of each other.
Fix it by composing each transformation with an existing one (starting with an
empty translation).