pointer-tracking: remove unused extensions
pointer-input-rectangle and pointer-place-rubber-band-line* are not
used in the codebase and are franz-specific extensions. I'm removing
them for now. If someone feels like re-adding them I think that such
utilities should be put in climf package (clim-franz-compat) and kept
in Vendor/ toplevel directory.

dragging-drawing seems like a useful extension, but on top of not
being specified it is not even exported nor used.