erase-output-record: be more lax with regard to repaint
When errorp is NIL then redraw the record's region on stream. It might
be that we have replayed it without adding the record to the stream's
history and we need to clean up after ourself.

This is indeed a case when we use DRAGGING-OUTPUT where we erase
record which is not part of the stream's history (produced with
with-output-to-output-record).