drag-output-record: rewrite function to be more conformant
Most notably we do respect repaint argument and multiple-window
argument. We address numerous edge cases which were erroneous:

- erase was called on the new coordinates instead of last ones
- repaint was not respected
- redundant keyword arguments were supplied
- transformation arguments were not always correct
- pointer motion and button press handlers were intermingled

For more sophisticated usage user has to provide a feedback
function. It is insufficient for a robust drag-and-drop support
because it features only erase/draw actions but there is no way around
it without breaking existing code. We either update the specification
or provide an alternative interface. Programmers may use
tracking-pointer macro.