pointer-tracking: provide a valid multiple-window interpretation
Previous interpretation was very confusing. Windows were interpreted
as frames (not sheets), grab-pointer was called during the operation
and it didn't even really work across different frames. Moreover there
was embrionic implementation of different interpretation, that window
is a sheet, but it applied only to pointer-motion and pointer-release
clauses.

Our current interpretation is as follows:

- when argument multiple-window is NIL, then pointer is tracked only
  for the sheet on which it was invoked

- when argument multiple-window is T, then pointer is tracked on all
  sheets on which we can track (i.e having the same event queue)

No grab-pointer is used (so we are not pointer hogs for other
windows).

Also pointer-tracking for frames should not by default highlight
presentations. That's what we have the highlight argument for.