panes: re-fix :COMPUTE space requirements in stream panes
TODO can we use CHANGING-SPACE-REQUIREMENTS?

Regarding :COMPUTE space requirements in stream panes, the spec says

  29.4.1 CLIM Stream Pane Options

  CLIM application frames accept [...] layout pane options. The space
  requirement options (:width, :height, and so forth) can also take a
  size specification of :compute, which causes CLIM to run the display
  function for the pane, and make the pane large enough to hold the
  output of the display function.

In McCLIM, the history so far is as follows:

* 7e6f3429 introduced the following code to be executed after
  redipslaying a stream pane to address this requirement;

    (when (or (equal :compute (pane-user-width pane))
              ...)
      (change-space-requirements))

* 9a13fa55 changed this to

    (unless (or (equal :compute (pane-user-width pane))
                ...)
      (change-space-requirements))

  to avoid infinite recursion.

The second change causes CHANGE-SPACE-REQUIREMENTS to be called after
every redisplay for stream panes that do /not/ specify the :COMPUTE
space requirement anywhere.

One example that demonstrates how this is problematic is the text size
demo:
1) changing the value of the text editor gadget uses the value-changed
   callback to redisplay the font pane
2) the redisplay of the editor gadget (since the DREI substrate
   involves a stream pane) calls CHANGE-SPACE-REQUIREMENTS due to the
   logic shown above which triggers various redisplays in the frame
3) the font pane, also a stream, may trigger the same process after
   its redisplay
These unnecessary redisplays are slow and flickery.

This commit restores the logic added in 7e6f3429 which is a direct
translation of the spec paragraph and adds a dedicated mechanism for
preventing infinite loops.