distribute-event: rewrite method for basic-port
We reintroduce a clean-slate implementation of the distribute-event
method for basic-port. Differences from previous implementation:

- No auxlliary methods (so port may simply specialize on port)
- Corrected coordinate fixup (get-pointer-position was buggy)
- Comments throughout the code explaining what is being done
- Shorter function body
- Pressed sheet implementation (i.e for scrolling)
- Grabbed sheet implementation
- Focused sheet implementation (for keyboard input)
- Correctly synthetized enter/leave events for unmirrored sheets

Pressed sheet is activated when we press the pointer button until we
release it. This sheet receives duplicated motion and release
events.

Grabbed sheet hijacks all events received from the port (except
boundary events which are ignored). Sheet grabbing may be achieved
programmatically with a macro WITH-POINTER-GRABBED.

Focused sheet receives keyboard events (if we introduce more input
events which are not tied to the pointer it is very likely that they
will be distributed to the focused sheet too). When port has no
focused sheet then keyboard events are ignored.

When the pointer is clicked over a sheet an event of type
WINDOW-MANAGER-FOCUS-EVENT is dispatched to it. If sheet wants to take
a focus it should call (SETF PORT-KEYBOARD-INPUT-FOCUS) on itself. On
top of that TOP-LEVEL-SHEET-MIXIN preserves the keyboard focus from
the last time when the window was focused.

Enter and leave events must be synthetized for sheets without
mirrors. Carefully commented implementation is provided which avoids
generating spurious duplicated events (unlike the old one). A
clear (event hierarchy -wise) distinction is made between ungrab
enter/leave events and "ordinary" ones.