text-style-character-width: return advance width (not the glyph's)
text-style-width calls text-style-character-width. I expect that
text-style-width (similar to text-style-height) will return a width of
the character cell box, not a width of the glyph's bounding box.

Consider drawing a grid for fixed font (so each character falls inside
cell), if we do not declare text-style-width to return the
advance-width creating such grid is impossible, because each glyph has
different width while advance-width's are the same.

Note that this does not change font-text-extent which returns correct
values for a bounding box and the minimal rectangle. This change makes
text-style-width and text-style-height more consistent.