clx ports: %realize-mirror: remove obsolete specializations
- border-pane specialization didn't request enter and leave
events (and other events which are normally delivered to sheets)

This is problematic because parts of some sheet hierarchies are not
mirrored and in that case it is border-pane (if it is mirrored) who
gets the events which need to be dispatched elsewhere.

- menu-button-pane specialization didn't request pointer-motion
events. We speculate that this could be for performance reasons.

- clim-stream-pane specialization did request additionally
pointer-motion-hint events.

That allows x11 server not sending all motion events but only
essential ones (before leaving the sheet, when button changes and upon
explicit query). We want to operate on all motion events hence this
specialization is removed.