clx ports: remove obsolete border handling bits
We speculate that originally drawing borders in McCLIM's CLX backend
was performed by clx while now it is performed in handle-repaint.

border-width argument was responsible for a border width drawn by
x11. It was not used thus it is removed.

border argument being 0 meant a black pixel. Since border-width was
always 0 it didn't matter.