clouseau: fix NOTE-OBJECT-OCCURRENCE
Without this change, when a place got a new value after the state
object for the old value recorded multiple occurrences but the new
value causes NOTE-OBJECT-OCCURRENCE to return immediately, the
previously recorded occurrences were not reset by the default
NOTE-OBJECT-OCCURRENCE method. This led to attempts to highlight
invalid combinations of occurrences and ultimately errors.

To avoid this problem, reset occurrences in a new :AROUND method
instead of the default method.