geometry: unbounded regions: provide tidy implementation
- Implementation for +nowhere+ and +everywhere+ is straightforward so
to unclutter the source code we move it to a single place.

- "General" methods specialized on region have been changed to
specialize on a bounding-rectangle to avoid ambiguity. For instance

   (defmethod region-difference ((a rectangle) (b region))
     (make-instance 'standard-region-difference a b))

would supersede a method which gives the correct answer for unbounded
region. b argument should have a specialization on bounding-rectangle.

- region-exclusive-or is demoted from a protocol function to an
utility. It is not used anywhere else and has only a single method.