tests: regions: ellipse: add few tests and switch to macrolet
macrolet instead of flet assures, that the original expression is
being printed in fiveam failure list.

Tests of region-contains-position-p when a distance of the point is
approximately 1 or less will likely fail because we factor an
additional epsilon for ellipse tests. Without that clx rendering of
rotated ellipses doesn't work correctly.