design, regions: sort out dependency issues regarding {no,every}where
Before this change, the compile and load order of design.lisp and
regions.lisp was not explicitly specified and loading design.lisp
first lead to a warning due to +IDENTITY-TRANSFORMATION+ not being
defined yet.

Generally speaking, the geometry module (including regions.lisp) is
more fundamental and should be compiled and loaded first. The only
obstacle were +{NO,EVERY}WHERE+, {NO,EVERY}WHERE-MIXIN which were
defined in design.lisp but initialized in regions.lisp. To allow
compiling and loading regions.lisp first (and avoid the definition and
initialization of +{NO,EVERY}WHERE+ in distinct places),
+{NO,EVERY}WHERE+ and {NO,EVERY}WHERE-MIXIN are now defined in
regions.lisp.