extended-streams: move files to a separate module
The only changes are file locations and asd definition. If we get rid
of Drei eventually, so we may have a single system definition, then
encapsulating-stream, stream-input and pointer-tracking will be part
of the "extended-input" (having also presentations) while this module
should be renamed to "extended-output" and have all formatting-* files.