text-selection: only steal events when actually selecting
Without this change, instances of TEXT-SELECTION-MIXIN rerouted
certain pointer events to EOS/SHIFT-{RELEASE,DRAG} as soon as the
shift modifier was active. This behavior was wrong because it did not
consider whether a text selection activity was in progress or not
(which is indicated by the DRAGGING-P slot of the pane and thus easily
checked).

fixes #961