geometry: region-intersection: use macrolet to reduce gf calls
region-intersection operation is symmetric so (f a b) === (f b a),
we've exploited that fact in a few places to reduce amount of
code. That introduced indirection by calling the generic function with
swapped arguments what /potentially/ could cause noticeable
performance regression. This commit inlines body in both cases by
means of macrolet.