Listener: Add command COM-SHOW-CLASS-DOCUMENTATION.
It can often be useful to be able to quickly get the documentation of
a class. This command takes a class name and prints the documentation
string if one is avalible. It also properly handles cases where either
the class does not exist or the documentation string does not exist.