frames: supply default methods generate-panes, find-pane-for-frame
Standard allows defining application frames without :pane(s) option,
so generate-panes should have a default method. For now we'll settle
on the interactor (to make commands useful).

Implement default method for find-pane-for-frame and use it in
adopt-frame method.

Cosmetic fix: don't bind *application-frame* in generated
generate-panes method because it is bound already by
with-look-and-feel-realization (so it is basically a no-op).

Fixes #934.