examples: re-enable "radar" pane in gadget-test.lisp
The timer event mechanism seems to work fine despite what the comments
claim. Why not exercise it?

There actually was something wrong, though. But the problem was in the
demo itself: The POINTS slot of the RADAR-PANE class was initialized
with a list literal which was mutated in the HANDLE-EVENT method.