stream-input: separate input-buffer from event-queue
We remove an invalid assumption that input-buffer is same thing as the
event-queue for standard-extended-input-stream. To do that we split
initargs and slots between standard-sheet-input-mixin and the internal
input-kernel-mixin (along with appropriate accessors).

CLIM after this change works as before except that SEIS signals an
unbound-slot condition when we try to read the STREAM-INPUT-BUFFER.

- add :type vector to the input-buffer slots
- untie input-buffer and event-queue
- remove stream-input-buffer references from sheet-input
- replace stream-input-buffer with sheet-event-queue where appropriate
- minor refactor of some code passages (stylistic)