stream-input: introduce INPUT-STREAM-KERNEL class
For easier review it has been introduced as a separate
commit. STANDARD-INPUT-STREAM and STANDARD-EXTENDED-INPUT-STREAM will
be build on top of it. Parts of the extended input stream and the
input editing stream protocol are implemented for this base class.

Extended Input Stream Protocol:

- stream-input-wait
- stream-listen
- stream-process-gesture
- stream-read-gesture
- stream-unread-gesture
- stream-clear-input
- stream-input-buffer, (setf stream-input-buffer)
- stream-rescanning-p

Input Editing Stream Protocol:

- stream-input-buffer, (setf stream-input-buffer)
- stream-scan-pointer, (setf stream-scan-opinter)
- stream-insertion-pointer, (setf stream-insertion-pointer)

Additionally it introduces its own protocol functions:

- stream-append-gesture
- stream-gesture-available-p

INPUT-STREAM-KERNEL class is meant to implement interaction between
the event queue (which is the sheet input protocol comain) and the
input buffer. This class is not directly mentioned in the spec but is
hinted in STANDARD-INPUT-STREAM and STANDARD-EXTENDED-INPUT-STREAM
class descriptions.

The gist of its inner workings is as follows:

- STREAM-INPUT-WAIT operation calls event-listen-or-wait and handles
  events available in the queue until input-buffer has available
  input. STREAM-LISTEN is just a trampoline to STREAM-INPUT-WAIT.

- all other methods which need access to stream gestures operate
  directly on the input-buffer. STREAM-READ-GESTURE invokes
  STREAM-INPUT-WAIT.