examples: remove invalid method from a stream-test demo
with-input-ediing macro redraws the interactor line, so unless the
input buffer is in transition state this will echo nothing. Example of
such behavior is a #\backspace character being in the input-buffer
waiting to be processed only when stream-read-gesture is called. This
test behave fine until now because we don't have real internal buffers
for standard-extended-input-stream objects.