pdf: allow using the graft with different units
Using the keyword :UNITS in WITH-OUTPUT-TO-PDF-STREAM, it is possible
to use the graft with different units in the PDF backend.

To calculate the right SHEET-NATIVE-TRANSFORMATION for the graft we
don't use SHEET-TRANSFORMATION of the graft which has no meaning for
the CLIM Specification because the graft has no parent, but we
introduce GRAFT-UNITS-TRANSFORMATION and
GRAFT-ORIENTATION-TRANSFORMATION functions.