pdf: using graft with different units
Using the keyword :UNITS in WITH-OUTPUT-TO-PDF-STREAM  its possible
to use graft with different units in pdf backend.

To calculate the right SHEET-NATIVE-TRANSFORMATION for the graft we
don't use SHEET-TRANSFORMATION of the graft that has no meaning for
the CLIM Specification because the graft has no parent, but we
introduce GRAFT-UNITS-TRANSFORMATION and
GRAFT-ORIENTATION-TRANSFORMATION functions.