postscript: use the graft in postscript backend
With this commit the postscript backend use the graft as already does
the pdf backend. The modification introduced in the previous commits
for pdf backend are introduced also for postscript backend.

In particular:
- the graft is created.
- the graft adopts the clim-postscript-stream.
- only the graft is mirrored and the mirror is the output file-stream
- the SHEET-NATIVE-REGION is setted according to DEVICE-TYPE of the port (:a4,
:a3, :letter) and it is +EVERYWHERE+ region for :eps.
- the SHEET-NATIVE-TRANSFORMATION of the graft is setted according to the
orientation and the units of the graft, using the new functions:
GRAFT-ORIENTATION-TRANSFORMATION and GRAFT-UNITS-TRANSFORMATION.
- function POSTSCRIPT-STREAM-FILE-STREAM is removed because it can be
substituted with the clim spec function SHEET-MIRROR.
- function POSTSCRIPT-MEDIUM-FILE-STREAM is removed because it can be
substituted with the clim spec function MEDIUM-DRAWABLE.
- in the postsctipt output the postscript header comment "%%Orientation"
is removed otherwise when it is set to "Landscape" it rotates the page but
also the coordinate system of 90 degree, and therefore it is necessary
to change the graft transformation according to page
orientation. Instead between Landscape and Portrait orientation I swap
the width with the height of the page.
- some tests are added.