pointer-tracking: synthesize events when the modifier state changes
This is primarily intended to help presentation drag and drop which
sometimes has to recompute applicable translators after modifier state
changes.

That said, other uses of TRACKING-POINTER may benefit as well since
the behavior of the :POINTER-MOTION clause is now slightly more
consistent: the clause receives events containing a modifier state but
was previously not invoked again when this modifier state changed.