In file clim-core/commands.lisp add an :initform nil for the docuemntation file of %menu-item.  This is necessary to avoid an error if the mouse moves over a command menu item
In file clim-core/gadgets/menu.lisp

- Introduce a new presentation type command-menu-item to be used in command menus diplayed in
appllication-frames.
- add a translator from the command-menu-item to command
- use the new presentation-type when displaying command table menu

These changes are necessary to fix a bug describbed  in #955, and folloed up with pull request #957.
The problem is that if a command-menu-item takes arguments, when the mouse moves over the item an error will
signaled rather than querying for the rest of the arguments;  this is because the documentation is not
initialized.  The  approach is to change the object (to the full command-menu-item rather than just a field of it)
and the presentation type used to present items in the command menu, and then to provide a
presentation-translator to command.  This causes the documentation in the pointer-documentation pane to just
show the command name.

The second change is to the accept method for command.   It wraps the call to the command-processor in
a with-input-context form, allowing it to distingish between a menu-click and a typed command.  This then
lets it invoke the partial-command-parser for a menu item with arguments that was clicked on.