Fix completion for logical pathnames
This fixes completions for all types of pathnames.
This addresses the same issues as 954 which has been closed.

Currently if the pathname begins with ~/ (meaning start from my home directory)
the pathname will not complete;

Also If the pathname is a logical pathname and there
is a file in the directory that the pathname resolves to and that file's name
includes a character that is illegal for a logical pathname then the pathname
will not complete.

The original filename-completer function is now broken into 3 steps:

1) Given the string to be complered, figure out what directory is
implied.  This considers both logical pathnames and pathnames relative to the
home directory.

2) In that directory collect all the files whose names begin with the string.
If the pathname is a logical pathname then filter out all files whose names
include characters that are illegal in logical pathnames (i.e. everything but
alpha numerics and dash).

3) Given the list of candidates from step 2 call the completer

This can be tested interactively (e.g. in the lisp-listener) by creating a
directory  and a logical host translating to it:

(progn
  (ensure-directories-exist "~/foobar/bar/")
  (setf (logical-pathname-translations "foo")
	  '(("**;*.*"	"~/foobar/**/*.*")))

  (with-open-file (f "~/foobar/bar/baz_bar.lisp" :direction :output :if-exists :supersede :if-does-not-exist :create ))
  (with-open-file (f "~/foobar/bar/baz-bar.lisp" :direction :output :if-exists :supersede :if-does-not-exist :create)))

1. (accept 'pathname) and typing ~/foobar/b followed by tab should complete to ~/foobar/bar
2. (accept 'pathname) and typing foo:b followed by tab should complete to foo:bar;
3. (accept 'pathname) and typing foo:bar;baz followed by tab should complete foo:bar;baz-bar.lisp