tests: line-breaks.random: relax the test for a newline character
New line characters are part of the text and line breaking algorithm
does not "remove them", that is there are no two markers: line start
and line end – only where the line should be broken. At least
according to unicode.  With this behavior in mind it is a text
renderer responsibility to ignore said control characters. Compare
with "Newline Function (NLF)" in UAX#14[1].

[1] https://www.unicode.org/reports/tr14/tr14-39.html