text-formatting: do not assume that the stream is a pane
standard-page-layout needs to access the page region (which is
initialized in a slot-unbound method for %page-region). When it is a
pane, then we want to have the viewport, otherwise it is the sheet
region. When the argument is a pane, then call the pane-viewport,
otherwise use the stream (or when pane-viewport returns NIL).