clx: fix SYNTHESIZE-POINTER-MOTION-EVENT for single mirroring
SYNTHESIZE-POINTER-MOTION-EVENT assumed that the pointer sheet of the
port has a direct mirror. With this change, when there is no direct
mirror, get the closest mirrored ancestor and use the native
transformation to translate from mirror coordinates into the
coordinate system of the pointer sheet.

Also a few cosmetic changes.