commands: rename ADD-{ACTUAL- -> }PRESENTATION-TRANSLATOR-TO-COMMAND-TABLE
The description in 27.5 Presentation Translator Utilities says the
function takes a translator name and a command table designator but
that doesn't make sense since the name could only be looked up in the
supplied command table to which it should be added.

Update our copy of the specification such that the translator
parameter is not a name but the translator object.