presentation-translators: add translator: fix the hash table key
simple-type-translators are keyed with the presentation-type-name and
in one place it was addressed with a full from-type value. In this
very function we use the presentation-type-name to remove it, similar
in the function remove-translator, so it is clearly a typo.

With the previous code adding a translator with a compound type like
`(or integer string)` resulted in an inaccessible translator.