presentation-translators: find-presentation-translators: fix it!
FIND-PRESENTATION-TRANSLATORS was defined to only look for translators
which had the same presentation metaclass as the from-type argument.
This assumption doesn't hold if we account for translators which work
on "meta" presentation types like OR, for instance:

    #<TRANSLATOR (or string integer) - command>
    (find-pesentation-translators 'integer 'command 'some-command-table)

With this change we iterate over all translators and check with
STUPID-SUBTYPEP whether the FROM-TYPE can be used with the
translator. This change renders the SIMPLE-TYPE-TRANSLATORS obsolete,
hence the slot, its accessors and updates are removed.