tests: add some tests to presentation-translators
Cases ALL-6 and ALL-7 do not pass, but they should.

The type (member "a" "b") is a subtype of a type (or string real)
The type (member "dan" 3) is a subtype of a type (or string real)

That is due to a defect in the PRESENTATION-SUBTYPEP. It might be so
that the defect is justified, see the comment above the function.