presentations: add a sanity check for the presentation's type
The presentation can't be NIL (which is a subtype of all other types),
for once the specification says, that the type NIL is useful only for
a wildcard translators (for to-type) - that makes also sense for the
from-type translators, however an existing presentation musn't have
the type NIL, because otherwise it "matches" every translator in
existance.

Such mistake was made in the with-input-editing call for accept-1,
wher sensitizer-type was specified to be NIL (instead of NULL).

Error in the initialize-instance :after method for the
presentation-mixin when the presentation type is NIl.